requirements-render.txt
numpy<2
pillow
trimesh
imageio
tqdm
pyrender==0.1.45
PyOpenGL==3.1.0
pyglet==1.5.27
accelerate
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
scikit-learn
gpustat
nvitop
diffusers
transformers
einops
huggingface_hub
opencv-python
omegaconf
scikit-image
peft
jaxtyping
typeguard
matplotlib
imageio-ffmpeg
deepspeed
wandb[media]
colormaps
hydra-core
objaverse
